changes in config

+++ Front/src/app/components/config/customer-list/customer-list.component.ts
@@ -6,6 +6,7 @@ import { CustomerViewModel } from '../../../models/CustomerModel';
 import { LoaderComponent } from '../../shared/loader/loader.component';
 import { CommonModule } from '@angular/common';
 import { ToastrService } from 'ngx-toastr';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-customer-list',
@@ -17,12 +18,13 @@ import { ToastrService } from 'ngx-toastr';
 export class CustomerListComponent {
   CustomerList: CustomerViewModel[] = [];
   loader: boolean = false;
+  DeletePopUp: boolean = false;
   constructor(
     private fb: FormBuilder,
     private router: Router,
     private customerService: CustomerService,
     private toastr: ToastrService
-  ) {}
+  ) { }
   ngOnInit(): void {
     this.ShowAllCustomers();
   }
@@ -39,15 +41,32 @@ export class CustomerListComponent {
   }
 
   DeleteCustomer(customerId: number) {
-    this.loader = true;
-    this.customerService.Delete(customerId).subscribe((result) => {
-      if (result.wasSuccessful) {
-        this.loader = false;
-        this.toastr.info(result.statusMessage, 'Cliente eliminado');
+
+    this.DeletePopUp = true
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esto borrara todos los datos previamente almacenados en la tabla.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#FFA500 ',
+      cancelButtonColor: '#6c757d',
+      confirmButtonText: 'Sí, Eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.loader = true;
+        this.customerService.Delete(customerId).subscribe((result) => {
+          if (result.wasSuccessful) {
+            this.loader = false;
+            this.toastr.info(result.statusMessage, 'Cliente eliminado');
+          } else {
+            this.toastr.error(result.statusMessage, 'error');
+          }
+          this.ShowAllCustomers();
+        });
       } else {
-        this.toastr.error(result.statusMessage, 'error');
+        this.DeletePopUp = false;
       }
-      this.ShowAllCustomers();
     });
   }
 }
