+++ Front/src/app/components/config/products/list-products/list-products.component.ts
@@ -2,10 +2,11 @@ import { ToastrService } from 'ngx-toastr';
 import { ProductModel } from '../../../../models/ProductModel';
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { ActivatedRoute, RouterLink} from '@angular/router';
+import { ActivatedRoute, RouterLink } from '@angular/router';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { ProductsService } from '../../../../services/products.service';
 import { ProductPriceInnerComponent } from "../../../admin/product-price-inner/product-price-inner.component";
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-list-products',
@@ -16,20 +17,21 @@ import { ProductPriceInnerComponent } from "../../../admin/product-price-inner/p
 })
 export class ListProductsComponent {
   listProducts: ProductModel[] = [];
-  subProductsList: ProductModel [] =[];
+  subProductsList: ProductModel[] = [];
   loading: boolean = false;
   expandedArea: boolean = true;
   id: number;
   producttochangeprice: number = -1;
+  DeletePopUp: boolean = false;
   constructor(
     private productService: ProductsService,
     private toastr: ToastrService,
     private aRoute: ActivatedRoute,
   ) {
     this.id = Number(aRoute.snapshot.paramMap.get('id'));
-   }
+  }
 
-  ngOnInit(id: number){
+  ngOnInit(id: number) {
     this.GetMain();
   };
   ChangePrice(productid: number) {
@@ -37,42 +39,62 @@ export class ListProductsComponent {
   }
   GetMain() {
     this.loading = true;
-    this.productService.GetAll().subscribe(result =>{
+    this.productService.GetAll().subscribe(result => {
       if (result.wasSuccessful == true) {
         this.listProducts = result.data;
-        this.loading= false;
+        this.loading = false;
       } else {
       }
     })
   };
 
-  GetByParentid(id: number){
+  GetByParentid(id: number) {
     this.loading
-    this.productService.GetByParentId(id).subscribe(result=> {
-      if(result.wasSuccessful == true){
+    this.productService.GetByParentId(id).subscribe(result => {
+      if (result.wasSuccessful == true) {
         this.subProductsList = result.data;
       }
     })
   }
 
-  Delete(id: number){
-    this.loading = true;
-    this.productService.Delete(id).subscribe(result  => {
-      if(result.wasSuccessful == true){
-        this.toastr.info(`El producto ${result.data.name} fue eliminado con exito`, `Producto Eliminado.`)
-      }else {
-        this.toastr.error('El producto no fue eliminado.', 'Error.');
+  Delete(id: number) {
+    this.DeletePopUp = true
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: 'Esto borrara todos los datos previamente almacenados en la tabla.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#FFA500 ',
+      cancelButtonColor: '#6c757d',
+      confirmButtonText: 'Sí, Eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.loading = true;
+        this.productService.Delete(id).subscribe(result => {
+          if (result.wasSuccessful == true) {
+            this.toastr.info(`El producto ${result.data.name} fue eliminado con exito`, `Producto Eliminado.`)
+          } else {
+            this.toastr.error('El producto no fue eliminado.', 'Error.');
+          }
+          this.GetMain();
+        })
+      } else {
+        this.DeletePopUp = false;
       }
-    this.GetMain();
-   })
+    });
+
+
+
+
   }
 
-  ExpandArea(id: number){
+  ExpandArea(id: number) {
     this.loading = true;
 
 
-    this.productService.GetByParentId(id).subscribe(result =>{
-      if(result.wasSuccessful == true){
+    this.productService.GetByParentId(id).subscribe(result => {
+      if (result.wasSuccessful == true) {
         this.expandedArea = true;
         this.subProductsList = result.data;
       }
